reafctor(run test): move helper to seperate module
